changes in database examples

+++ ch15/hasql/Statements.hs
@@ -14,7 +14,7 @@ import Data.Text (Text)
 
 import FilmInfo.Data
 
-fiFromTuple :: (Int64, Text, Text, Int32, Text)
+fiFromTuple :: (Int32, Text, Text, Int32, Text)
                 -> FilmInfo
 fiFromTuple (i, t, d, l, r) = FilmInfo {
     filmId = FilmId i
@@ -24,16 +24,16 @@ fiFromTuple (i, t, d, l, r) = FilmInfo {
   , rating = toMaybeRating r
   }
 
-fromFilmId :: FilmId -> Int64
+fromFilmId :: FilmId -> Int32
 fromFilmId (FilmId a) = a
 
-fromCatId :: CatId -> Int64
+fromCatId :: CatId -> Int32
 fromCatId (CatId a) = a
 
 allFilms :: Statement () (Vector FilmInfo)
 allFilms = rmap (fmap fiFromTuple)
   [TH.vectorStatement|
-     SELECT film_id :: int8, title :: text,
+     SELECT film_id :: int4, title :: text,
             description :: text,
             length :: int4, rating :: text
      FROM film
@@ -48,7 +48,7 @@ countFilms =
 findFilm :: Statement Text (Maybe FilmInfo)
 findFilm = rmap (fmap fiFromTuple)
   [TH.maybeStatement|
-     SELECT film_id :: int8, title :: text,
+     SELECT film_id :: int4, title :: text,
             description :: text,
             length :: int4, rating :: text
      FROM film
@@ -58,7 +58,7 @@ findFilm = rmap (fmap fiFromTuple)
 filmsLonger :: Statement Int32 (Vector FilmInfo)
 filmsLonger = rmap (fmap fiFromTuple)
   [TH.vectorStatement|
-     SELECT film_id :: int8, title :: text,
+     SELECT film_id :: int4, title :: text,
             description :: text,
             length :: int4, rating :: text
      FROM film
@@ -84,34 +84,34 @@ setRating = lmap (first' fromRating)
 filmIdByTitle :: Statement Text (Maybe FilmId)
 filmIdByTitle = rmap (fmap FilmId)
   [TH.maybeStatement|
-     SELECT film_id::int8 FROM film WHERE title=$1::text
+     SELECT film_id::int4 FROM film WHERE title=$1::text
   |]
 
 catIdByName :: Statement Text (Maybe CatId)
 catIdByName = rmap (fmap CatId)
   [TH.maybeStatement|
-     SELECT category_id::int8 FROM category WHERE name=$1::text
+     SELECT category_id::int4 FROM category WHERE name=$1::text
   |]
 
 newCategory :: Statement Text CatId
 newCategory = rmap CatId
   [TH.singletonStatement|
     INSERT INTO category (name) VALUES ($1::text)
-    RETURNING category_id::int8
+    RETURNING category_id::int4
   |]
 
 isAssigned :: Statement (CatId, FilmId) Bool
 isAssigned = lmap (bimap fromCatId fromFilmId) $ rmap isJust
   [TH.maybeStatement|
-     SELECT category_id::int8 FROM film_category
-     WHERE category_id=$1::int8 AND film_id=$2::int8
+     SELECT category_id::int4 FROM film_category
+     WHERE category_id=$1::int4 AND film_id=$2::int4
   |]
 
 assignCategory :: Statement (CatId, FilmId) Int64
 assignCategory = lmap (bimap fromCatId fromFilmId) $
   [TH.rowsAffectedStatement|
      INSERT INTO film_category (category_id, film_id)
-     VALUES ($1::int8, $2::int8)
+     VALUES ($1::int4, $2::int4)
   |]
 
 unassignCategory :: Statement (Text, Text) Int64
